# cache __TYPE_MAP and init submodules