--- a/.gitignore
+++ b/.gitignore
@@ -80,3 +80,4 @@ _version.py
 
 .core_typemap_version
 core_typemap.pkl
+core_nscatalog.pkl