--- a/.gitignore
+++ b/.gitignore
@@ -80,4 +80,3 @@ _version.py
 
 .core_typemap_version
 core_typemap.pkl
-core_nscatalog.pkl