--- a/pyproject.toml
+++ b/pyproject.toml
@@ -98,7 +98,7 @@ exclude = [
   "__pycache__",
   "build/",
   "dist/",
-  "src/nwb-schema",
+  "src/pynwb/nwb-schema",
   "docs/source/conf.py",
   "src/pynwb/_due.py",
   "test.py" # remove when pytest comes along