--- a/tests/unit/test_import.py
+++ b/tests/unit/test_import.py
@@ -11,6 +11,7 @@ try:
 except ImportError:
     from importlib_resources import files
 
+import pynwb
 from pynwb.testing import TestCase
 
 class TestPyNWBSubmoduleClone(TestCase):
@@ -23,9 +24,7 @@ class TestPyNWBSubmoduleClone(TestCase):
 
         # remove cache files if they exist
         self.typemap_cache = Path(files('pynwb') / 'core_typemap.pkl')
-        self.nscatalog_cache = Path(files('pynwb') / 'core_nscatalog.pkl')
         self.typemap_cache.unlink(missing_ok=True)
-        self.nscatalog_cache.unlink(missing_ok=True)
 
 
     def tearDown(self) -> None:
@@ -40,11 +39,9 @@ class TestPyNWBSubmoduleClone(TestCase):
         we should try and import it
         """
         assert not self.repo_dir.exists()
-        assert not self.nscatalog_cache.exists()
         assert not self.typemap_cache.exists()
 
         with self.assertWarns(Warning) as clone_warning:
-            import pynwb
             importlib.reload(pynwb)
 
             # there can be other warnings, but we should get at least one
@@ -53,7 +50,6 @@ class TestPyNWBSubmoduleClone(TestCase):
             assert any(['initializing submodules' in msg for msg in messages])
 
         assert self.repo_dir.exists()
-        assert self.nscatalog_cache.exists()
         assert self.typemap_cache.exists()
 
         # we should also get the namespaces correctly too
@@ -72,5 +68,4 @@ class TestPyNWBImportCache(TestCase):
         The global __NS_CATALOG and the one contained within the global __TYPE_MAP
         should be the same object after importing
         """
-        import pynwb
-        assert id(pynwb.__TYPE_MAP.namespace_catalog) == id(pynwb.__NS_CATALOG)
+        assert id(getattr(pynwb, '__TYPE_MAP').namespace_catalog) == id(getattr(pynwb, '__NS_CATALOG'))