@sneakers-the-rat 2024-07-10
(lmao i must hear the story here)
<img width="586" alt="Screenshot 2024-07-09 at 10 54 53 PM" src="https://github.com/NeurodataWithoutBorders/pynwb/assets/12961499/2af5fc40-1fd3-46f3-88f3-3660ccd3b30b">