@sneakers-the-rat 2024-07-11
OK i tested whether importing all the submodules first before pickling sped up the import further, and it doesn't. That version is here: https://github.com/sneakers-the-rat/pynwb/commit/0a1514ee29ea6aba372ea1c946624aa9372b9701

The profiling results on my computer (obviously variable) are:
- Before caching: 4.33s
- After caching: 1.05s
- Improvement: 3.28s (75%)

The reason that including the imported modules doesn't improve the speed much if any is that the main contributor to the slow imports is that a) there are import side effects, b) those import side effects include parsing yaml c) parsing yaml is slow. The module imports take very little time because the yaml is already parsed by the time they are imported, so imo those should remain uncached because unpickling objects with declaration-time side effects (the registration decorator) is super wonky and fragile, u can see the hack i had to do in that branch.

There are three remaining opportunities for future import perf improvements:
- `hdmf` basically duplicates the import side effects of `pynwb`, so a similar thing could be done there for an additional ~300ms off
- the `YAMLSpecReader` in `hdmf` uses the slower Python-based yaml parser because of the `pure=True` kwarg here: https://github.com/hdmf-dev/hdmf/blob/dfb1df79bce3e34c52af7996429c3f561d96390a/src/hdmf/spec/namespace.py#L203 
  as far as I'm aware all the NWB schemas are yaml 1.1 compliant, so i am not sure there is a reason for doing that. Removing that and using the C-based loader saves an additional 150ms after caching.
- having all these import side effects seems to have lent itself to having a very big import tree - importing `pynwb` imports 765 modules (numpy and scipy, two very heavy packages, import 114 and 124 respectively). Fixing this would be a more involved refactoring where logic would have to be moved out of `__init__.py` files, import side effects would have to be turned into on demand loading, and then the package level imports could be pruned down. The remaining ~700ms of time is just the normal amount of time it takes to import 765 modules.

I'm marking this as ready for review without adding tests just to see if this approach would be acceptable, and if so then i'll add tests for the submodule cloning, caching, and cache invalidation.