@sneakers-the-rat 2024-07-16
Huh, so the test that's failing is this one https://github.com/NeurodataWithoutBorders/pynwb/blob/570fb3b396b3c2c03b4abb6af23d8d76b1c26586/tests/back_compat/test_import_structure.py#L6

which makes sense because I removed `__resources` from the module-level scope. I figured that would be OK since it is a dunder attribute, a "really private, really don't try and access these from outside the scope" name, and so i'm sort of surprised that they would be in that API consistency test? 

Lmk what would be better, updating that test (i think that none of those dunder attributes should be in there?) or making it so `__resources` is back in the module scope. I think that might just be a code style thing, in general i try and avoid using module level globals, esp in this case where it's a single use, quickly computed thing, but it would really be no problem to put it back in module globals.