@sneakers-the-rat 2024-07-16
OK moved `__resources` back to the module scope, and I added a simple test for the submodule cloning - incidentally that requires one to remove the cache files, so it implicitly functions as a test for the creation but not the use of the cache files. I am not sure the best way to test that the cache file is actually used - i could do something like emit a warning there, but that seems noisy, or i could add some kind of env var flag, but that also seems inelegant, so lmk what y'all would like there.

during testing i discovered that one also needs to cache `__NS_CATALOG`, which makes sense because it looks like one wraps the other. I restored the global version from the one that is saved in the cached `__TYPE_MAP` so we only need one cache file. Added a very simple test that their `id` matches.

One more thing is that while linting i noticed that the ruff `excludes` path points to the wrong directory for `nwb-schema` so i adjusted that too.