@sneakers-the-rat 2024-07-17
gee for a simple thing this sure does involve a lot of fighting lol

OK so i had written some tests for this but since we're talking about import side effects testing them is hard.
- it turns out that `__TYPE_MAP` also stores the generated classes, so a test for import effects of the top-level `pynwb` fails because on reimport the references to the objects are regenerated, so then `isinstance()` and equality checks fail and those seem to cascade through hdmf in a way i can't really tell where the problem comes in. so anyway that breaks the rest of the tests
- calling the `_load_core_namespace` manually to trigger a creation of the cache file again doesn't work because by the time the import tests run there has been an example extension test run, but it doesn't clean up after itself in the type map so there is an unpickleable class (because it doesn't have a namespace, it's (`TestExtension.test_lab_meta.<locals>.MyTestMetaData`), whomp whomp.
- the package is not installed in editable mode from tox, so it's no longer in a git repo when the tests run, so testing cloning submodules fails there (but succeeds when run from the repo)

fortunately....
- the tests all run and pass when using the cached TYPE_MAP, so those import/symbol identity inconsistencies are not true for the pickled file, presumably because unpickling runs the code they come from and the names are bound then.
- the submodule cloning code has the same error path - if the submodule directory doesn't exist, throw error. so it's at worse net neutral in terms of bugs (actually a little bit better - if you delete the schema directory the type map still loads correctly),

so I just scrapped the tests because this is ultimately a very simple change and i did not budget this much time to this very simple change. 

if that's not gonna cut it feel free to close this, just trying to do a little QoL contribs while i work with this package so not mission critical, fine for me to just have it as a fork.