@stephprince 2024-07-18
Hi @sneakers-the-rat, thanks for all the work and profiling! Really appreciate the walkthrough.

@rly said he can take a look at this so I'll let him review in detail.