@rly 2024-09-17
I made a few small changes. Looks good! Thanks so much for this PR and catching the slowdown with `get_type_map`. I'll look into that next.